Change source structure, hide internals
