--- packages/core/resolve-redux/src/constants.ts
@@ -1,3 +1,0 @@
-export const aggregateVersionsMap = '@@resolve/aggregateVersionsMap'
-export const lastTimestampMap = '@@resolve/lastTimestampMap'
-export const connectorMetaMap = '@@resolve/connectorMetaMap'
