--- packages/core/resolve-redux/src/create_jwt_reducer.ts
@@ -1,16 +1,0 @@
-import { UPDATE_JWT, LOGOUT } from './action-types'
-
-const createJwtReducer = () => (state = {}, action: any): any => {
-  switch (action.type) {
-    case UPDATE_JWT: {
-      return action.jwt
-    }
-    case LOGOUT: {
-      return {}
-    }
-    default:
-      return state
-  }
-}
-
-export default createJwtReducer
