--- packages/core/resolve-redux/src/empty_saga.ts
@@ -1,3 +1,0 @@
-const emptySaga = function*() {}
-
-export default emptySaga
