--- packages/core/resolve-redux/src/empty_subscribe_adapter.ts
@@ -1,19 +1,0 @@
-const emptySubscribeAdapter = () => ({
-  init() {
-    return Promise.resolve()
-  },
-  subscribeToTopics() {
-    return Promise.resolve()
-  },
-  unsubscribeFromTopics() {
-    return Promise.resolve()
-  },
-  close() {
-    return Promise.resolve()
-  },
-  isConnected: (): boolean => true
-})
-
-emptySubscribeAdapter.adapterName = 'empty'
-
-export default emptySubscribeAdapter
