+++ packages/core/resolve-redux/src/index.ts
@@ -1,64 +1,44 @@
-import createJwtReducer from './create_jwt_reducer'
+import AppContainer from './app-container'
+import createStore from './create-store'
+
+import connectStaticBasedUrls from './urls/connect-static-based-urls'
+import connectRootBasedUrls from './urls/connect-root-based-urls'
 import connectReadModel from './read-model/connect-read-model'
 import connectViewModel from './view-model/connect-view-model'
-import connectResolveAdvanced from './connect_resolve_advanced'
-import createApi from './create_api'
-import createConnectionManager from './create_connection_manager'
-import createSagaManager from './create_saga_manager'
-import getHash from './get-hash'
-import getRootBasedUrl from './get_root_based_url'
-import getStaticBasedUrl from './get_static_based_url'
-import connectStaticBasedUrls from './connect_static_based_urls'
-import connectRootBasedUrls from './connect_root_based_urls'
-import { FetchError, HttpError } from './create_api'
-import { Provider, Consumer } from './resolve_context'
-import Providers from './providers'
-import AppContainer from './app_container'
-import deserializeInitialState from './deserialize_initial_state'
-import createStore from './create_store'
-import { sendAggregateAction } from './command/send-aggregate-action'
-import getOrigin from './get_origin'
 
+import { sendAggregateAction } from './command/send-aggregate-action'
 import { useReduxReadModel } from './read-model/use-redux-read-model'
 import { useReduxReadModelSelector } from './read-model/use-redux-read-model-selector'
 import { useReduxCommand } from './command/use-redux-command'
 import { useReduxViewModel } from './view-model/use-redux-view-model'
 
-import * as actionTypes from './action-types'
+import getOrigin from './internal/get-origin'
+import { Provider, Consumer } from './internal/resolve-context'
+import Providers from './internal/providers'
+import * as actionTypes from './internal/action-types'
 
 import { ResultStatus } from './types'
 
 const internal = {
-  actionTypes
+  actionTypes,
+  getOrigin,
+  Provider,
+  Consumer,
+  Providers
 }
 
 export {
-  createJwtReducer,
+  AppContainer,
+  createStore,
+  sendAggregateAction,
   connectViewModel,
   connectReadModel,
-  connectResolveAdvanced,
   connectStaticBasedUrls,
   connectRootBasedUrls,
-  createApi,
-  createConnectionManager,
-  createSagaManager,
-  getHash,
-  getRootBasedUrl,
-  getStaticBasedUrl,
-  Provider,
-  Consumer,
-  FetchError,
-  HttpError,
-  AppContainer,
-  Providers,
-  createStore,
-  deserializeInitialState,
-  sendAggregateAction,
-  getOrigin,
-  internal,
   useReduxReadModel,
   useReduxReadModelSelector,
   useReduxCommand,
   useReduxViewModel,
-  ResultStatus
+  ResultStatus,
+  internal
 }
