+++ packages/core/resolve-redux/src/internal/actions.ts
@@ -1,0 +1,6 @@
+import { HOT_MODULE_REPLACEMENT } from '../internal/action-types'
+
+export const hotModuleReplacement = (hotModuleReplacementId: any) => ({
+  type: HOT_MODULE_REPLACEMENT,
+  hotModuleReplacementId
+})
