+++ packages/core/resolve-redux/src/internal/create-saga-manager.ts
