+++ packages/core/resolve-redux/src/internal/get-origin.ts
