--- packages/core/resolve-redux/src/logout_saga.ts
@@ -1,10 +1,0 @@
-const logoutSaga = function*({
-  jwtProvider
-}: {
-  type: string
-  jwtProvider: any
-}) {
-  yield jwtProvider.set('')
-}
-
-export default logoutSaga
