+++ packages/core/resolve-redux/src/read-model/connect-read-model-saga.ts
@@ -6,7 +6,7 @@ import {
   queryReadModelRequest,
   queryReadModelSuccess
 } from './actions'
-import { DISCONNECT_READMODEL } from '../action-types'
+import { DISCONNECT_READMODEL } from '../internal/action-types'
 
 import { RootSagaArgs } from '../types'
