--- packages/core/resolve-redux/src/sync_jwt_provider_with_store.ts
@@ -1,23 +1,0 @@
-import jwtDecode from 'jwt-decode'
-import stableStringify from 'json-stable-stringify'
-import { updateJwt } from './actions'
-
-const syncJwtProviderWithStore = async (
-  jwtProvider: any,
-  store: any
-): Promise<void> => {
-  if (jwtProvider) {
-    const token = await jwtProvider.get()
-
-    let jwt = store.getState().jwt
-    try {
-      jwt = jwtDecode(token)
-    } catch (err) {}
-
-    if (stableStringify(store.getState().jwt) !== stableStringify(jwt)) {
-      store.dispatch(updateJwt(jwt))
-    }
-  }
-}
-
-export default syncJwtProviderWithStore
