+++ packages/core/resolve-redux/src/view-model/view-model-reducer.ts
@@ -1,9 +1,9 @@
-import getHash from '../get-hash'
+import getHash from '../internal/get-hash'
 import setEntry from 'lodash.set'
 import unsetEntry from 'lodash.unset'
 import getByPath from 'lodash.get'
 
-import { DROP_VIEWMODEL_STATE, VIEWMODEL_STATE_UPDATE } from '../action-types'
+import { DROP_VIEWMODEL_STATE, VIEWMODEL_STATE_UPDATE } from '../internal/action-types'
 import {
   ViewModelResultEntry,
   ViewModelResultMapByName,
